requirements.txt
pandas>=2.0
numpy>=1.24
pyyaml>=6.0
matplotlib>=3.8
pyarrow>=14.0
geopandas>=0.14